Start to add WIP of a gerrit plugin
This plugin will have a similar purpose to the existing
gerritbot but instead of being a singular application it
will be part of the composable framework that we can
combine together to build a good (single) openstack bot.

Currently this does not configure or understand the
older configuration format, and it behaves and formats
messages slightly differently, but the overall concept
is the same.

Change-Id: I3fa12fd765bace9e66c788c19940741adefd139c